ui/volumes/nodes:
Move the URL Sync logic for both Volumes and Nodes tables to its custom hook function to avoid a big block of code redundancy.

Refs: #2919
